requirement.txt
asgiref==3.5.2
Django==4.1.3
django-environ==0.9.0
django-filter==22.1
django-versatileimagefield==2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-flex-fields==1.0.0
mysqlclient==2.1.1
Pillow==9.3.0
PyJWT==2.6.0
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7
